[IMP] in pec_messages field to determintae if message is input oroutput writable only for admin for old pec mail [FIX] view for mail pec [FIX] permission for pec_user [IMP] new module to mange pec in hierarchical structures folders
