+++ l10n_it_pec_messages/model/mail_message.py
@@ -25,97 +25,6 @@
 
 from openerp.osv import fields, orm
 
-#----------------------------------------------------------
-# Categories
-#----------------------------------------------------------
-class message_category(orm.Model):
-
-    def name_get(self, cr, uid, ids, context=None):
-        if isinstance(ids, (list, tuple)) and not len(ids):
-            return []
-        if isinstance(ids, (long, int)):
-            ids = [ids]
-        reads = self.read(cr, uid, ids, ['name','parent_id'], context=context)
-        res = []
-        for record in reads:
-            name = record['name']
-            if record['parent_id']:
-                name = record['parent_id'][1]+' / '+name
-            res.append((record['id'], name))
-        return res
-
-    def _name_get_fnc(self, cr, uid, ids, prop, unknow_none, context=None):
-        res = self.name_get(cr, uid, ids, context=context)
-        return dict(res)
-
-    def _get_def_server(self, cr, uid, ids, *a):
-        res = {}
-        user_pool=self.pool.get('res.users')
-        user=user_pool.browse(cr, uid, [uid])[0]
-        return user.allowed_server_ids[0].id
-
-
-    _name = "message.category"
-    _description = "Message Category"
-    _columns = {
-        'name': fields.char(
-            'Name', size=64, required=True, translate=True, select=True),
-        'complete_name': fields.function(
-            _name_get_fnc, type="char", string='Name'),
-        'parent_id': fields.many2one(
-            'message.category','Parent Category',
-            select=True, ondelete='cascade'),
-        'child_id': fields.one2many(
-            'message.category', 'parent_id', string='Child Categories'),
-        'sequence': fields.integer(
-            'Sequence', select=True,
-            help=(
-                "Gives the sequence order when displaying"
-                "a list of product categories."
-            )
-        ),
-        'type': fields.selection(
-            [('view','View'), ('normal','Normal')], 'Category Type',
-            help=(
-                "A category of the view type is a virtual category "
-                "that can be used as the parent of another category "
-                "to create a hierarchical structure."
-            )
-        ),
-        'parent_left': fields.integer('Left Parent', select=1),
-        'parent_right': fields.integer('Right Parent', select=1),
-        'server_id': fields.many2one(
-            'fetchmail.server', 'Server Pec', readonly=True),
-    }
-
-
-    _defaults = {
-        'server_id': _get_def_server,
-        'type' : lambda *a : 'normal',
-    }
-
-    _parent_name = "parent_id"
-    _parent_store = True
-    _parent_order = 'sequence, name'
-    _order = 'parent_left'
-
-    def _check_recursion(self, cr, uid, ids, context=None):
-        level = 100
-        while len(ids):
-            cr.execute('select distinct parent_id from message_category where id IN %s',(tuple(ids),))
-            ids = filter(None, map(lambda x:x[0], cr.fetchall()))
-            if not level:
-                return False
-            level -= 1
-        return True
-
-    _constraints = [
-        (_check_recursion, 'Error ! You cannot create recursive categories.', ['parent_id'])
-    ]
-    def child_get(self, cr, uid, ids):
-        return [ids]
-
-
 
 class MailMessage(orm.Model):
     _inherit = "mail.message"
@@ -136,6 +45,10 @@ class MailMessage(orm.Model):
             _get_out_server, type='many2one',
             relation='ir.mail_server',
             string='Related Outgoing Server'),
+        'direction': fields.selection([
+            ('in', 'in'),
+            ('out', 'out'),
+            ], 'Mail direction'),
         'pec_type': fields.selection([
             ('posta-certificata', 'Pec Mail'),
             ('accettazione', 'Reception'),
@@ -193,10 +106,10 @@ class MailMessage(orm.Model):
         'main_msg_virus_message_ids': fields.one2many(
             'mail.message', 'virus_message_id',
             'Main messages', readonly=True),
-        'folder_id': fields.many2one(
-            'message.category','Folder', required=False,
-            change_default=True,
-            help="Select category for the current product"),
+    }
+
+    _defaults = {
+        'direction': 'in'
     }
 
     def _search(
