+++ l10n_it_pec_messages/wizard/mail_compose_message_view.xml
@@ -8,10 +8,10 @@
                 <form string="Compose Email" version="7.0">
                     <group>
                         <!-- truly invisible fields for control and options -->
-                        <field name="composition_mode" invisible="1"/>
-                        <field name="model" invisible="1"/>
-                        <field name="res_id" invisible="1"/>
-                        <field name="parent_id" invisible="1"/>
+                        <field name="composition_mode" invisible="1" />
+                        <field name="model" invisible="1" />
+                        <field name="res_id" invisible="1" />
+                        <field name="parent_id" invisible="1" context="{'pec_messages': True}"/>
                         <!-- visible wizard -->
                         <field name="server_id" />
                         <label for="partner_ids" string="Recipients"
@@ -25,15 +25,14 @@
                                            attrs="{'invisible':[('model', '=', False)]}"/>
                                     and
                             </span>
-                            <field name="partner_ids" widget="many2many_tags" placeholder="Add contacts to notify..."
-                                       context="{'show_pec_email':True}"/>
+                            <field name="partner_ids" widget="many2many_tags" placeholder="Add contacts to notify..."/>
                         </div>
-                        <field name="subject" placeholder="Subject..."/>
+                        <field name="subject" placeholder="Subject..." />
                     </group>
                     <field name="body"/>
-                    <field name="attachment_ids" widget="many2many_binary"/>
+                    <field name="attachment_ids" widget="many2many_binary" />
                     <footer>
-                        <button string="Send" name="send_mail" type="object" class="oe_highlight"/>
+                        <button string="Send" name="send_mail" type="object" class="oe_highlight" />
                         or
                         <button string="Cancel" class="oe_link" special="cancel" />
                     </footer>
@@ -47,7 +46,7 @@
             <field name="type">ir.actions.act_window</field>
             <field name="view_type">form</field>
             <field name="view_mode">form</field>
-            <field name="context" >{'new_pec_mail': True}</field>
+            <field name="context" >{'new_pec_mail': True,'pec_messages': True}</field>
             <field name="target">new</field>
         </record>
 
@@ -57,13 +56,14 @@
             <field name="type">ir.actions.act_window</field>
             <field name="view_type">form</field>
             <field name="view_mode">form</field>
-            <field name="context" >{'new_pec_mail': True,'reply_pec':True}</field>
+            <field name="context" >{'new_pec_mail': True,'reply_pec':True,'pec_messages': True}</field>
             <field name="target">new</field>
         </record>
 
         <record id="action_write_pec_wizard" model="ir.actions.act_window.view">
             <field name="sequence" eval="4"/>
             <field name="view_mode">form</field>
+            <field name="context" >{'new_pec_mail': True,'reply_pec':True,'pec_messages': True}</field>
             <field name="view_id" ref="email_compose_message_wizard_form_pec"/>
             <field name="act_window_id" ref="action_email_pec_compose_message_wizard"/>
         </record>
