+++ l10n_it_pec_messages_folders/security/ir.model.access.csv
@@ -1,0 +1,2 @@
+id,name,model_id:id,group_id:id,perm_read,perm_write,perm_create,perm_unlink
+access_mail_message_folders,access_mail_message_folders,model_message_category,base.group_pec,1,1,1,0
