+++ l10n_it_pec_messages_folders/view/mail_view.xml
@@ -1,0 +1,146 @@
+<?xml version="1.0" encoding="UTF-8"?>
+<openerp>
+<data>
+
+        <record model='ir.ui.view' id='message_pec_folders_form'>
+            <field name="name">message.pec.folder.form</field>
+            <field name="model">mail.message</field>
+            <field name="inherit_id" ref="l10n_it_pec_messages.message_pec_form"/>
+            <field name="priority">100</field>
+            <field name="arch" type="xml">
+                    <field name="author_id" position="after" >
+                        <field name="folder_id"  />
+                    </field>
+                    <field name="child_ids" position="attributes">
+                             <attribute name="context">{
+                                              'form_view_ref' : 'l10n_it_pec_messages_folders.message_pec_folders_form',
+                                              'tree_view_ref' : 'l10n_it_pec_messages_folders.message_pec_folders_tree',
+                                              'pec_messages': True}
+                                </attribute>
+                      </field>
+            </field>
+        </record>
+
+        <record model='ir.ui.view' id='message_pec_folders_tree'>
+            <field name="name">message.pec.tree</field>
+            <field name="model">mail.message</field>
+            <field name="inherit_id" ref="l10n_it_pec_messages.message_pec_tree"/>
+            <field name="priority">100</field>
+            <field name="arch" type="xml">
+                    <field name="email_from"  position="before">
+                        <field name="folder_id"  invisible="1"/>
+                    </field>
+            </field>
+        </record>
+
+        <record id="pec_message_folders_search_view" model="ir.ui.view">
+            <field name="name">message.search.form</field>
+            <field name="model">mail.message</field>
+            <field name="inherit_id" ref="l10n_it_pec_messages.pec_message_search_view"/>
+            <field name="arch" type="xml">
+               <field name="date"  position="after">
+                    <separator/>
+                   <field name="folder_id"/>
+                   <group  expand='0' string='Group by...'>
+                       <filter string='Folder' icon="terp-stock_symbol-selection" domain="[]" context="{'group_by' : 'folder_id'}"/>
+                   </group>
+               </field>
+            </field>
+        </record>
+
+
+<!--
+        <record model="ir.actions.act_window.view" id="message_pec_act_tree">
+            <field name="sequence" eval="1"/>
+            <field name="view_mode">tree</field>
+            <field name="view_id" ref="message_pec_folder_tree"/>
+            <field name="act_window_id" ref="action_mail_pec"/>
+        </record>
+
+        <record model="ir.actions.act_window.view" id="message_pec_act_form">
+            <field name="sequence" eval="2"/>
+            <field name="view_mode">form</field>
+            <field name="view_id" ref="message_pec_folder_t"/>
+            <field name="act_window_id" ref="action_mail_pec"/>
+        </record>
+-->
+
+
+        <record id="message_category_search_view" model="ir.ui.view">
+            <field name="name">message.category.search</field>
+            <field name="model">message.category</field>
+            <field name="arch" type="xml">
+                <search string="Message Categories">
+                    <field name="name" string="Message Folders"/>
+                    <field name="parent_id"/>
+                </search>
+            </field>
+        </record>
+
+        <record id="message_category_form_view" model="ir.ui.view">
+            <field name="name">message.category.form</field>
+            <field name="model">message.category</field>
+            <field name="arch" type="xml">
+                <form string="Product Folders" version="7.0">
+                    <sheet>
+                        <div class="oe_title">
+                            <label for="name" class="oe_edit_only"/>
+                            <h1>
+                                <field name="name"/>
+                            </h1>
+                        </div>
+                        <group>
+                            <group name="parent" col="4">
+                                <field name="parent_id"/>
+                                <field name="type"/>
+                                <field name="server_id" />
+                            </group>
+                        </group>
+                    </sheet>
+                </form>
+            </field>
+        </record>
+
+        <record id="message_category_list_view" model="ir.ui.view">
+            <field name="name">message.category.list</field>
+            <field name="model">message.category</field>
+            <field name="priority">1</field>
+            <field name="arch" type="xml">
+                <tree string="Message Folders">
+                    <field name="complete_name"/>
+                </tree>
+            </field>
+        </record>
+
+        <record id="message_category_tree_view" model="ir.ui.view">
+            <field name="name">message.category.tree</field>
+            <field name="model">message.category</field>
+            <field name="field_parent">child_id</field>
+            <field name="arch" type="xml">
+                <tree toolbar="True" string="Message Folders">
+                    <field name="name"/>
+                </tree>
+            </field>
+        </record>
+
+        <record id="message_category_action_form" model="ir.actions.act_window">
+            <field name="name">Pec Mail Folders</field>
+            <field name="type">ir.actions.act_window</field>
+            <field name="res_model">message.category</field>
+            <field name="view_type">form</field>
+            <field name="domain">[
+                  ('server_id.pec','=',True),
+                  ('server_id.user_ids','in',uid)]</field>
+            <field name="search_view_id" ref="message_category_search_view"/>
+            <field name="view_id" ref="message_category_list_view"/>
+        </record>
+
+        <record id="config_mail_pec" model="ir.ui.menu">
+            <field name="name">Config PEC Folders</field>
+            <field name="action" ref="message_category_action_form"/>
+            <field name="parent_id" ref="l10n_it_pec_messages.mail_pec"/>
+            <field name="groups_id" eval="[(4, ref('base.group_pec'))]"/>
+        </record>
+
+</data>
+</openerp>
